follow the code rules in the README
